# Bloggify-swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Bloggify",
    "contact": {}
  },
  "host": "localhost:5000",
  "basePath": "/api/user",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/blogs": {
      "get": {
        "summary": "Get All Blogs",
        "tags": [
          "Users"
        ],
        "operationId": "GetAllBlogs",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "cat",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetAllBlogs-SuccessCategoryWise"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "status": 1,
                "message": "Blogs Found Success.",
                "data": [
                  {
                    "_id": "6197d41950d3cfb6e44feda5",
                    "category_id": "6198bd633e59225d6bbef84d",
                    "image": "https://bloggify.s3.amazonaws.com/cdfe9408-b688-4521-83f0-a1da6a41e4fd.png",
                    "image_alt": "immune-system",
                    "title": "Make your immune system more strong | Boost your Immunity",
                    "author_id": "6197d24dabb407fc6a11b1aa",
                    "content": "How does the Immune System Work?\nWhen the body senses foreign substances (called antigens), the system works to acknowledge the antigens and acquire to eliminate them.\n\nLymphocytes are triggered to make antibodies. These specialized proteins lock onto specific antigens. The antibodies stay in a person's body. That way, if the system encounters that antigen again, the antibodies are ready to do their job. That's why someone who gets sick with a disease, like chickenpox, usually won't get sick from it again.\n\nAntibodies can neutralize toxins (poisonous or damaging substances) created by totally different organisms and activate a bunch of proteins referred to as complement that is a section of the system. Complement helps kill microorganisms, viruses, or infected cells. These specialized cells and components of the system offer body protection against disease. This protection is termed as immunity.\nFever is an immune system response\n\nHumans have 3 forms of immunity — innate, adaptive, and passive.\n\nInnate immunity :\nEveryone seems to be born with innate (or natural) immunity, a kind of general protection.\n\nAdaptive immunity :\nAdaptive (or active) immunity develops throughout our lives. We tend to develop adaptive immunity once we're exposed to diseases or once we're immunized against them with vaccines.\n\nPassive immunity :\nPassive immunity is \"borrowed\" from another source and it lasts for a quick time.",
                    "status": "published",
                    "date": "2021-11-19T16:43:05.545Z",
                    "createdAt": "2021-11-19T16:43:05.570Z",
                    "updatedAt": "2021-11-20T12:12:52.209Z"
                  }
                ]
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "1940"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"794-44OrEnLryRJlWb/bppdBPj4nrjg\""
              },
              "Date": {
                "type": "string",
                "default": "Sat, 20 Nov 2021 14:30:17 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Data2": {
      "title": "Data2",
      "example": {
        "category_id": "6197cd59e28a5eb64301b2c8",
        "image": "https://bloggify.s3.amazonaws.com/90dee1de-9a03-493a-90b2-c37cd4079306.jpeg",
        "image_alt": "90dee1de-9a03-493a-90b2-c37cd4079306",
        "title": "Why should you become a full stack developer?",
        "author_id": "6199039f2b88fabc684ab32f",
        "content": "Let’s get this cleared first. Here are the top 3 reasons why you should consider becoming a full stack developer -\n\n1. Building unique proprietary software products\nEvery company/developer follows their own tech stack while building a product (as every product has its own requirements and the best set of technology is chosen accordingly).\n\nThe use of such diverse technologies makes the code base of the project complex and this eventually creates a unique code for a particular product. Companies adopting full stack development will have this edge.",
        "status": "published",
        "date": "2021-11-20T14:23:37.117Z",
        "_id": "619904e92b88fabc684ab332",
        "createdAt": "2021-11-20T14:23:37.130Z",
        "updatedAt": "2021-11-20T14:23:37.130Z"
      },
      "type": "object",
      "properties": {
        "category_id": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "image_alt": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "author_id": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "_id": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      },
      "required": [
        "category_id",
        "image",
        "image_alt",
        "title",
        "author_id",
        "content",
        "status",
        "date",
        "_id",
        "createdAt",
        "updatedAt"
      ]
    },
...
    "GetAllBlogs-SuccessCategoryWise": {
      "title": "GetAllBlogs-SuccessCategoryWise",
      "example": {
        "status": 1,
        "message": "Blogs Found Success.",
        "data": [
          {
            "_id": "6197d41950d3cfb6e44feda5",
            "category_id": "6198bd633e59225d6bbef84d",
            "image": "https://bloggify.s3.amazonaws.com/cdfe9408-b688-4521-83f0-a1da6a41e4fd.png",
            "image_alt": "immune-system",
            "title": "Make your immune system more strong | Boost your Immunity",
            "author_id": "6197d24dabb407fc6a11b1aa",
            "content": "How does the Immune System Work?\nWhen the body senses foreign substances (called antigens), the system works to acknowledge the antigens and acquire to eliminate them.\n\nLymphocytes are triggered to make antibodies. These specialized proteins lock onto specific antigens. The antibodies stay in a person's body. That way, if the system encounters that antigen again, the antibodies are ready to do their job. That's why someone who gets sick with a disease, like chickenpox, usually won't get sick from it again.\n\nAntibodies can neutralize toxins (poisonous or damaging substances) created by totally different organisms and activate a bunch of proteins referred to as complement that is a section of the system. Complement helps kill microorganisms, viruses, or infected cells. These specialized cells and components of the system offer body protection against disease. This protection is termed as immunity.\nFever is an immune system response\n\nHumans have 3 forms of immunity — innate, adaptive, and passive.\n\nInnate immunity :\nEveryone seems to be born with innate (or natural) immunity, a kind of general protection.\n\nAdaptive immunity :\nAdaptive (or active) immunity develops throughout our lives. We tend to develop adaptive immunity once we're exposed to diseases or once we're immunized against them with vaccines.\n\nPassive immunity :\nPassive immunity is \"borrowed\" from another source and it lasts for a quick time.",
            "status": "published",
            "date": "2021-11-19T16:43:05.545Z",
            "createdAt": "2021-11-19T16:43:05.570Z",
            "updatedAt": "2021-11-20T12:12:52.209Z"
          }
        ]
      },
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Data2"
          }
        }
      },
      "required": [
        "status",
        "message",
        "data"
      ]
    },
...
  },
...
}